+++ generate_patch.c
@@ -16,8 +16,8 @@ int main(int argc, char** argv) {
 
     fseek(old_fp, 0L, SEEK_END);
     fseek(new_fp, 0L, SEEK_END);
-    long old_len = ftell(old_fp);
-    long new_len = ftell(new_fp);
+    uint32_t old_len = ftell(old_fp);
+    uint32_t new_len = ftell(new_fp);
     if(old_len != new_len) {
         printf("files do not match in length. simple_bin_patch does not support shifting reigons of memory.\n\n");
         fclose(old_fp); fclose(new_fp);
@@ -28,21 +28,21 @@ int main(int argc, char** argv) {
 
     FILE* diff_fp = fopen(argv[3], "wb");
 
-    byte* old_buf = (byte*)malloc(old_len);
-    byte* new_buf = (byte*)malloc(new_len);
+    uint8_t* old_buf = (uint8_t*)malloc(old_len);
+    uint8_t* new_buf = (uint8_t*)malloc(new_len);
 
-    fread(old_buf, sizeof(byte), old_len, old_fp);
-    fread(new_buf, sizeof(byte), new_len, new_fp);
+    fread(old_buf, sizeof(uint8_t), old_len, old_fp);
+    fread(new_buf, sizeof(uint8_t), new_len, new_fp);
 
     // the worst case scenario is if every other byte is the same and
     // every other byte is different. 1 byte for the opcode, 4 for the
     // length, and one for the byte. hence the *6. in those cases, this
     // algorithm shouldn't be used.
-    byte* diff_buf = (byte*)calloc(old_len * 6, sizeof(byte));
+    uint8_t* diff_buf = (uint8_t*)calloc(old_len * 6, sizeof(uint8_t));
 
-    long diff_len = generate_patch(old_buf, new_buf, diff_buf, old_len);
+    uint32_t diff_len = generate_patch(old_buf, new_buf, diff_buf, old_len);
 
-    fwrite(diff_buf, sizeof(byte), diff_len, diff_fp);
+    fwrite(diff_buf, sizeof(uint8_t), diff_len, diff_fp);
 
     fclose(old_fp); fclose(new_fp); fclose(diff_fp);
     free(old_buf);  free(new_buf);  free(diff_buf);
